+++ backend/schedulerd/check_watcher_test.go
@@ -1,3 +1,5 @@
+// +build integration
+
 package schedulerd
 
 import (